services/orion-dream/requirements.txt
fastapi==0.111.*
uvicorn[standard]==0.30.*
redis==5.0.*
sqlalchemy==2.0.*
psycopg2-binary==2.9.*
pydantic==2.7.*
pydantic-settings==2.5.2
httpx==0.27.*
chromadb==0.4.24
rdflib==7.0.0
numpy==1.26.*
requests==2.32.*
python-dotenv==1.0.*
scikit-learn==1.4.2
